오늘 의존성작업. Configuration 클래스 수정해줌. 메인페이지 프레임작업. 헤더, 푸터 분리
