+++ src/main/java/com/sds/foodfit/controller/MainController.java
@@ -7,7 +7,13 @@ import org.springframework.web.bind.annotation.GetMapping;
 public class MainController {
   	@GetMapping("/")
 	public String getMain() {
-		return "index";
+		return "main/index";
+		
+	}
+  	
+  	@GetMapping("/demo")
+	public String goDemo() {
+		return "demo/input_form";
 		
 	}
