+++ src/main/java/com/sds/foodfit/domain/FoodRecommendResult.java
@@ -1,5 +1,8 @@
 package com.sds.foodfit.domain;
 
+import lombok.Data;
+
+@Data
 public class FoodRecommendResult {
 	public int food_recommend_res_idx;
 	public String recommend_res_foodnm;
