+++ src/main/java/com/sds/foodfit/domain/Member.java
@@ -4,5 +4,6 @@ import lombok.Data;
 
 @Data
 public class Member {
+	
 
 }
