+++ src/main/resources/application.properties
@@ -1,10 +1,19 @@
 spring.application.name=FoodFit
 
 
-#mysql 구성
+#mysql 구성 (회원DB에 사용)
 spring.datasource.url=jdbc:mysql://localhost:3306/food
 spring.datasource.username=root
 spring.datasource.password=1234
 spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
 spring.jpa.hibernate.ddl-auto=update
 
+#로거 설정
+logging.level.com.sds.foodfit=DEBUG
+
+#타임리프 설정
+spring.thymeleaf.enabled=true
+spring.thymeleaf.cache=false
+spring.thymeleaf.prefix=classpath:/templates/
+spring.thymeleaf.suffix=.html
+spring.thymeleaf.encoding=UTF-8
