+++ src/main/resources/static/site/about.html
@@ -27,8 +27,8 @@
     <link rel="stylesheet" href="css/custom.css">
 
     <!--[if lt IE 9]>
-      <script src="https://oss.maxcdn.com/libs/html5shiv/3.7.0/html5shiv.js"></script>
-      <script src="https://oss.maxcdn.com/libs/respond.js/1.4.2/respond.min.js"></script>
+      <script src="site/https://oss.maxcdn.com/libs/html5shiv/3.7.0/html5shiv.js"></script>
+      <script src="site/https://oss.maxcdn.com/libs/respond.js/1.4.2/respond.min.js"></script>
     <![endif]-->
 
 </head>
@@ -39,7 +39,7 @@
 		<nav class="navbar navbar-expand-lg navbar-light bg-light">
 			<div class="container">
 				<a class="navbar-brand" href="index.html">
-					<img src="images/logo.png" alt="" />
+					<img src="site/images/logo.png" alt="" />
 				</a>
 				<button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbars-rs-food" aria-controls="navbars-rs-food" aria-expanded="false" aria-label="Toggle navigation">
 				  <span class="navbar-toggler-icon"></span>
@@ -89,7 +89,7 @@
 		<div class="container">
 			<div class="row">
 				<div class="col-lg-6 col-md-6">
-					<img src="images/about-img.jpg" alt="" class="img-fluid">
+					<img src="site/images/about-img.jpg" alt="" class="img-fluid">
 				</div>
 				<div class="col-lg-6 col-md-6 text-center">
 					<div class="inner-column">
@@ -138,7 +138,7 @@
 			<div class="row special-list">
 				<div class="col-lg-4 col-md-6 special-grid drinks">
 					<div class="gallery-single fix">
-						<img src="images/img-01.jpg" class="img-fluid" alt="Image">
+						<img src="site/images/img-01.jpg" class="img-fluid" alt="Image">
 						<div class="why-text">
 							<h4>Special Drinks 1</h4>
 							<p>Sed id magna vitae eros sagittis euismod.</p>
@@ -149,7 +149,7 @@
 				
 				<div class="col-lg-4 col-md-6 special-grid drinks">
 					<div class="gallery-single fix">
-						<img src="images/img-02.jpg" class="img-fluid" alt="Image">
+						<img src="site/images/img-02.jpg" class="img-fluid" alt="Image">
 						<div class="why-text">
 							<h4>Special Drinks 2</h4>
 							<p>Sed id magna vitae eros sagittis euismod.</p>
@@ -160,7 +160,7 @@
 				
 				<div class="col-lg-4 col-md-6 special-grid drinks">
 					<div class="gallery-single fix">
-						<img src="images/img-03.jpg" class="img-fluid" alt="Image">
+						<img src="site/images/img-03.jpg" class="img-fluid" alt="Image">
 						<div class="why-text">
 							<h4>Special Drinks 3</h4>
 							<p>Sed id magna vitae eros sagittis euismod.</p>
@@ -171,7 +171,7 @@
 				
 				<div class="col-lg-4 col-md-6 special-grid lunch">
 					<div class="gallery-single fix">
-						<img src="images/img-04.jpg" class="img-fluid" alt="Image">
+						<img src="site/images/img-04.jpg" class="img-fluid" alt="Image">
 						<div class="why-text">
 							<h4>Special Lunch 1</h4>
 							<p>Sed id magna vitae eros sagittis euismod.</p>
@@ -182,7 +182,7 @@
 				
 				<div class="col-lg-4 col-md-6 special-grid lunch">
 					<div class="gallery-single fix">
-						<img src="images/img-05.jpg" class="img-fluid" alt="Image">
+						<img src="site/images/img-05.jpg" class="img-fluid" alt="Image">
 						<div class="why-text">
 							<h4>Special Lunch 2</h4>
 							<p>Sed id magna vitae eros sagittis euismod.</p>
@@ -193,7 +193,7 @@
 				
 				<div class="col-lg-4 col-md-6 special-grid lunch">
 					<div class="gallery-single fix">
-						<img src="images/img-06.jpg" class="img-fluid" alt="Image">
+						<img src="site/images/img-06.jpg" class="img-fluid" alt="Image">
 						<div class="why-text">
 							<h4>Special Lunch 3</h4>
 							<p>Sed id magna vitae eros sagittis euismod.</p>
@@ -204,7 +204,7 @@
 				
 				<div class="col-lg-4 col-md-6 special-grid dinner">
 					<div class="gallery-single fix">
-						<img src="images/img-07.jpg" class="img-fluid" alt="Image">
+						<img src="site/images/img-07.jpg" class="img-fluid" alt="Image">
 						<div class="why-text">
 							<h4>Special Dinner 1</h4>
 							<p>Sed id magna vitae eros sagittis euismod.</p>
@@ -215,7 +215,7 @@
 				
 				<div class="col-lg-4 col-md-6 special-grid dinner">
 					<div class="gallery-single fix">
-						<img src="images/img-08.jpg" class="img-fluid" alt="Image">
+						<img src="site/images/img-08.jpg" class="img-fluid" alt="Image">
 						<div class="why-text">
 							<h4>Special Dinner 2</h4>
 							<p>Sed id magna vitae eros sagittis euismod.</p>
@@ -226,7 +226,7 @@
 				
 				<div class="col-lg-4 col-md-6 special-grid dinner">
 					<div class="gallery-single fix">
-						<img src="images/img-09.jpg" class="img-fluid" alt="Image">
+						<img src="site/images/img-09.jpg" class="img-fluid" alt="Image">
 						<div class="why-text">
 							<h4>Special Dinner 3</h4>
 							<p>Sed id magna vitae eros sagittis euismod.</p>
@@ -334,16 +334,16 @@
 	<a href="#" id="back-to-top" title="Back to top" style="display: none;">&uarr;</a>
 
 	<!-- ALL JS FILES -->
-	<script src="js/jquery-3.2.1.min.js"></script>
-	<script src="js/popper.min.js"></script>
-	<script src="js/bootstrap.min.js"></script>
+	<script src="site/js/jquery-3.2.1.min.js"></script>
+	<script src="site/js/popper.min.js"></script>
+	<script src="site/js/bootstrap.min.js"></script>
     <!-- ALL PLUGINS -->
-	<script src="js/jquery.superslides.min.js"></script>
-	<script src="js/images-loded.min.js"></script>
-	<script src="js/isotope.min.js"></script>
-	<script src="js/baguetteBox.min.js"></script>
-	<script src="js/form-validator.min.js"></script>
-    <script src="js/contact-form-script.js"></script>
-    <script src="js/custom.js"></script>
+	<script src="site/js/jquery.superslides.min.js"></script>
+	<script src="site/js/images-loded.min.js"></script>
+	<script src="site/js/isotope.min.js"></script>
+	<script src="site/js/baguetteBox.min.js"></script>
+	<script src="site/js/form-validator.min.js"></script>
+    <script src="site/js/contact-form-script.js"></script>
+    <script src="site/js/custom.js"></script>
 </body>
 </html>
