+++ src/main/resources/static/site/blog-details.html
@@ -27,8 +27,8 @@
     <link rel="stylesheet" href="css/custom.css">
 
     <!--[if lt IE 9]>
-      <script src="https://oss.maxcdn.com/libs/html5shiv/3.7.0/html5shiv.js"></script>
-      <script src="https://oss.maxcdn.com/libs/respond.js/1.4.2/respond.min.js"></script>
+      <script src="site/https://oss.maxcdn.com/libs/html5shiv/3.7.0/html5shiv.js"></script>
+      <script src="site/https://oss.maxcdn.com/libs/respond.js/1.4.2/respond.min.js"></script>
     <![endif]-->
 
 </head>
@@ -39,7 +39,7 @@
 		<nav class="navbar navbar-expand-lg navbar-light bg-light">
 			<div class="container">
 				<a class="navbar-brand" href="index.html">
-					<img src="images/logo.png" alt="" />
+					<img src="site/images/logo.png" alt="" />
 				</a>
 				<button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbars-rs-food" aria-controls="navbars-rs-food" aria-expanded="false" aria-label="Toggle navigation">
 				  <span class="navbar-toggler-icon"></span>
@@ -100,7 +100,7 @@
 					<div class="blog-inner-details-page">
 						<div class="blog-inner-box">
 							<div class="side-blog-img">
-								<img class="img-fluid" src="images/inner-blog-img.jpg" alt="">							
+								<img class="img-fluid" src="site/images/inner-blog-img.jpg" alt="">							
 								<div class="date-blog-up">
 									27 Fab
 								</div>
@@ -124,7 +124,7 @@
 							<h3>Comments</h3>
 							<div class="comment-item">
 								<div class="comment-item-left">
-									<img src="images/avt-img.jpg" alt="">
+									<img src="site/images/avt-img.jpg" alt="">
 								</div>
 								<div class="comment-item-right">
 									<div class="pull-left">
@@ -141,7 +141,7 @@
 							</div>
 							<div class="comment-item children">
 								<div class="comment-item-left">
-									<img src="images/avt-img.jpg" alt="">
+									<img src="site/images/avt-img.jpg" alt="">
 								</div>
 								<div class="comment-item-right">
 									<div class="pull-left">
@@ -158,7 +158,7 @@
 							</div>
 							<div class="comment-item">
 								<div class="comment-item-left">
-									<img src="images/avt-img.jpg" alt="">
+									<img src="site/images/avt-img.jpg" alt="">
 								</div>
 								<div class="comment-item-right">
 									<div class="pull-left">
@@ -225,7 +225,7 @@
 							<div class="recent-post-box">
 								<div class="recent-box-blog">
 									<div class="recent-img">
-										<img class="img-fluid" src="images/post-img-01.jpg" alt="">
+										<img class="img-fluid" src="site/images/post-img-01.jpg" alt="">
 									</div>
 									<div class="recent-info">
 										<ul>
@@ -238,7 +238,7 @@
 								</div>
 								<div class="recent-box-blog">
 									<div class="recent-img">
-										<img class="img-fluid" src="images/post-img-02.jpg" alt="">
+										<img class="img-fluid" src="site/images/post-img-02.jpg" alt="">
 									</div>
 									<div class="recent-info">
 										<ul>
@@ -251,7 +251,7 @@
 								</div>
 								<div class="recent-box-blog">
 									<div class="recent-img">
-										<img class="img-fluid" src="images/post-img-03.jpg" alt="">
+										<img class="img-fluid" src="site/images/post-img-03.jpg" alt="">
 									</div>
 									<div class="recent-info">
 										<ul>
@@ -264,7 +264,7 @@
 								</div>
 								<div class="recent-box-blog">
 									<div class="recent-img">
-										<img class="img-fluid" src="images/post-img-01.jpg" alt="">
+										<img class="img-fluid" src="site/images/post-img-01.jpg" alt="">
 									</div>
 									<div class="recent-info">
 										<ul>
@@ -390,17 +390,17 @@
 	<a href="#" id="back-to-top" title="Back to top" style="display: none;">&uarr;</a>
 
 	<!-- ALL JS FILES -->
-	<script src="js/jquery-3.2.1.min.js"></script>
-	<script src="js/popper.min.js"></script>
-	<script src="js/bootstrap.min.js"></script>
+	<script src="site/js/jquery-3.2.1.min.js"></script>
+	<script src="site/js/popper.min.js"></script>
+	<script src="site/js/bootstrap.min.js"></script>
     <!-- ALL PLUGINS -->
 	
-	<script src="js/jquery.superslides.min.js"></script>
-	<script src="js/images-loded.min.js"></script>
-	<script src="js/isotope.min.js"></script>
-	<script src="js/baguetteBox.min.js"></script>
-	<script src="js/form-validator.min.js"></script>
-    <script src="js/contact-form-script.js"></script>
-    <script src="js/custom.js"></script>
+	<script src="site/js/jquery.superslides.min.js"></script>
+	<script src="site/js/images-loded.min.js"></script>
+	<script src="site/js/isotope.min.js"></script>
+	<script src="site/js/baguetteBox.min.js"></script>
+	<script src="site/js/form-validator.min.js"></script>
+    <script src="site/js/contact-form-script.js"></script>
+    <script src="site/js/custom.js"></script>
 </body>
 </html>
