+++ src/main/resources/static/site/blog.html
@@ -27,8 +27,8 @@
     <link rel="stylesheet" href="css/custom.css">
 
     <!--[if lt IE 9]>
-      <script src="https://oss.maxcdn.com/libs/html5shiv/3.7.0/html5shiv.js"></script>
-      <script src="https://oss.maxcdn.com/libs/respond.js/1.4.2/respond.min.js"></script>
+      <script src="site/https://oss.maxcdn.com/libs/html5shiv/3.7.0/html5shiv.js"></script>
+      <script src="site/https://oss.maxcdn.com/libs/respond.js/1.4.2/respond.min.js"></script>
     <![endif]-->
 
 </head>
@@ -39,7 +39,7 @@
 		<nav class="navbar navbar-expand-lg navbar-light bg-light">
 			<div class="container">
 				<a class="navbar-brand" href="index.html">
-					<img src="images/logo.png" alt="" />
+					<img src="site/images/logo.png" alt="" />
 				</a>
 				<button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbars-rs-food" aria-controls="navbars-rs-food" aria-expanded="false" aria-label="Toggle navigation">
 				  <span class="navbar-toggler-icon"></span>
@@ -99,7 +99,7 @@
 				<div class="col-lg-4 col-md-6 col-12">
 					<div class="blog-box-inner">
 						<div class="blog-img-box">
-							<img class="img-fluid" src="images/blog-img-01.jpg" alt="">
+							<img class="img-fluid" src="site/images/blog-img-01.jpg" alt="">
 						</div>
 						<div class="blog-detail">
 							<h4>Duis feugiat neque sed dolor cursus.</h4>
@@ -117,7 +117,7 @@
 				<div class="col-lg-4 col-md-6 col-12">
 					<div class="blog-box-inner">
 						<div class="blog-img-box">
-							<img class="img-fluid" src="images/blog-img-02.jpg" alt="">
+							<img class="img-fluid" src="site/images/blog-img-02.jpg" alt="">
 						</div>
 						<div class="blog-detail">
 							<h4>Duis feugiat neque sed dolor cursus.</h4>
@@ -135,7 +135,7 @@
 				<div class="col-lg-4 col-md-6 col-12">
 					<div class="blog-box-inner">
 						<div class="blog-img-box">
-							<img class="img-fluid" src="images/blog-img-03.jpg" alt="">
+							<img class="img-fluid" src="site/images/blog-img-03.jpg" alt="">
 						</div>
 						<div class="blog-detail">
 							<h4>Duis feugiat neque sed dolor cursus.</h4>
@@ -153,7 +153,7 @@
 				<div class="col-lg-4 col-md-6 col-12">
 					<div class="blog-box-inner">
 						<div class="blog-img-box">
-							<img class="img-fluid" src="images/blog-img-04.jpg" alt="">
+							<img class="img-fluid" src="site/images/blog-img-04.jpg" alt="">
 						</div>
 						<div class="blog-detail">
 							<h4>Duis feugiat neque sed dolor cursus.</h4>
@@ -171,7 +171,7 @@
 				<div class="col-lg-4 col-md-6 col-12">
 					<div class="blog-box-inner">
 						<div class="blog-img-box">
-							<img class="img-fluid" src="images/blog-img-05.jpg" alt="">
+							<img class="img-fluid" src="site/images/blog-img-05.jpg" alt="">
 						</div>
 						<div class="blog-detail">
 							<h4>Duis feugiat neque sed dolor cursus.</h4>
@@ -189,7 +189,7 @@
 				<div class="col-lg-4 col-md-6 col-12">
 					<div class="blog-box-inner">
 						<div class="blog-img-box">
-							<img class="img-fluid" src="images/blog-img-06.jpg" alt="">
+							<img class="img-fluid" src="site/images/blog-img-06.jpg" alt="">
 						</div>
 						<div class="blog-detail">
 							<h4>Duis feugiat neque sed dolor cursus.</h4>
@@ -207,7 +207,7 @@
 				<div class="col-lg-4 col-md-6 col-12">
 					<div class="blog-box-inner">
 						<div class="blog-img-box">
-							<img class="img-fluid" src="images/blog-img-07.jpg" alt="">
+							<img class="img-fluid" src="site/images/blog-img-07.jpg" alt="">
 						</div>
 						<div class="blog-detail">
 							<h4>Duis feugiat neque sed dolor cursus.</h4>
@@ -225,7 +225,7 @@
 				<div class="col-lg-4 col-md-6 col-12">
 					<div class="blog-box-inner">
 						<div class="blog-img-box">
-							<img class="img-fluid" src="images/blog-img-08.jpg" alt="">
+							<img class="img-fluid" src="site/images/blog-img-08.jpg" alt="">
 						</div>
 						<div class="blog-detail">
 							<h4>Duis feugiat neque sed dolor cursus.</h4>
@@ -243,7 +243,7 @@
 				<div class="col-lg-4 col-md-6 col-12">
 					<div class="blog-box-inner">
 						<div class="blog-img-box">
-							<img class="img-fluid" src="images/blog-img-09.jpg" alt="">
+							<img class="img-fluid" src="site/images/blog-img-09.jpg" alt="">
 						</div>
 						<div class="blog-detail">
 							<h4>Duis feugiat neque sed dolor cursus.</h4>
@@ -357,17 +357,17 @@
 	<a href="#" id="back-to-top" title="Back to top" style="display: none;">&uarr;</a>
 
 	<!-- ALL JS FILES -->
-	<script src="js/jquery-3.2.1.min.js"></script>
-	<script src="js/popper.min.js"></script>
-	<script src="js/bootstrap.min.js"></script>
+	<script src="site/js/jquery-3.2.1.min.js"></script>
+	<script src="site/js/popper.min.js"></script>
+	<script src="site/js/bootstrap.min.js"></script>
     <!-- ALL PLUGINS -->
 	
-	<script src="js/jquery.superslides.min.js"></script>
-	<script src="js/images-loded.min.js"></script>
-	<script src="js/isotope.min.js"></script>
-	<script src="js/baguetteBox.min.js"></script>
-	<script src="js/form-validator.min.js"></script>
-    <script src="js/contact-form-script.js"></script>
-    <script src="js/custom.js"></script>
+	<script src="site/js/jquery.superslides.min.js"></script>
+	<script src="site/js/images-loded.min.js"></script>
+	<script src="site/js/isotope.min.js"></script>
+	<script src="site/js/baguetteBox.min.js"></script>
+	<script src="site/js/form-validator.min.js"></script>
+    <script src="site/js/contact-form-script.js"></script>
+    <script src="site/js/custom.js"></script>
 </body>
 </html>
