+++ src/main/resources/templates/fragments/footer.html
@@ -5,52 +5,77 @@
 			<div class="row">
 				<div class="col-lg-3 col-md-6">
 					<h3>About Us</h3>
-					<p>Integer cursus scelerisque ipsum id efficitur. Donec a dui fringilla, gravida lorem ac, semper magna. Aenean rhoncus ac lectus a interdum. Vivamus semper posuere dui, at ornare turpis ultrices sit amet. Nulla cursus lorem ut nisi porta, ac eleifend arcu ultrices.</p>
+					<p>Integer cursus scelerisque ipsum id efficitur. Donec a dui
+						fringilla, gravida lorem ac, semper magna. Aenean rhoncus ac
+						lectus a interdum. Vivamus semper posuere dui, at ornare turpis
+						ultrices sit amet. Nulla cursus lorem ut nisi porta, ac eleifend
+						arcu ultrices.</p>
 				</div>
 				<div class="col-lg-3 col-md-6">
 					<h3>Opening hours</h3>
-					<p><span class="text-color">Monday: </span>Closed</p>
-					<p><span class="text-color">Tue-Wed :</span> 9:Am - 10PM</p>
-					<p><span class="text-color">Thu-Fri :</span> 9:Am - 10PM</p>
-					<p><span class="text-color">Sat-Sun :</span> 5:PM - 10PM</p>
+					<p>
+						<span class="text-color">Monday: </span>Closed
+					</p>
+					<p>
+						<span class="text-color">Tue-Wed :</span> 9:Am - 10PM
+					</p>
+					<p>
+						<span class="text-color">Thu-Fri :</span> 9:Am - 10PM
+					</p>
+					<p>
+						<span class="text-color">Sat-Sun :</span> 5:PM - 10PM
+					</p>
 				</div>
 				<div class="col-lg-3 col-md-6">
 					<h3>Contact information</h3>
 					<p class="lead">Ipsum Street, Lorem Tower, MO, Columbia, 508000</p>
-					<p class="lead"><a href="#">+01 2000 800 9999</a></p>
-					<p><a href="#"> info@admin.com</a></p>
+					<p class="lead">
+						<a href="#">+01 2000 800 9999</a>
+					</p>
+					<p>
+						<a href="#"> info@admin.com</a>
+					</p>
 				</div>
 				<div class="col-lg-3 col-md-6">
 					<h3>Subscribe</h3>
 					<div class="subscribe_form">
 						<form class="subscribe_form">
-							<input name="EMAIL" id="subs-email" class="form_input" placeholder="Email Address..." type="email">
+							<input name="EMAIL" id="subs-email" class="form_input"
+								placeholder="Email Address..." type="email">
 							<button type="submit" class="submit">SUBSCRIBE</button>
 							<div class="clearfix"></div>
 						</form>
 					</div>
 					<ul class="list-inline f-social">
-						<li class="list-inline-item"><a href="#"><i class="fa fa-facebook" aria-hidden="true"></i></a></li>
-						<li class="list-inline-item"><a href="#"><i class="fa fa-twitter" aria-hidden="true"></i></a></li>
-						<li class="list-inline-item"><a href="#"><i class="fa fa-linkedin" aria-hidden="true"></i></a></li>
-						<li class="list-inline-item"><a href="#"><i class="fa fa-google-plus" aria-hidden="true"></i></a></li>
-						<li class="list-inline-item"><a href="#"><i class="fa fa-instagram" aria-hidden="true"></i></a></li>
+						<li class="list-inline-item"><a href="#"><i
+								class="fa fa-facebook" aria-hidden="true"></i></a></li>
+						<li class="list-inline-item"><a href="#"><i
+								class="fa fa-twitter" aria-hidden="true"></i></a></li>
+						<li class="list-inline-item"><a href="#"><i
+								class="fa fa-linkedin" aria-hidden="true"></i></a></li>
+						<li class="list-inline-item"><a href="#"><i
+								class="fa fa-google-plus" aria-hidden="true"></i></a></li>
+						<li class="list-inline-item"><a href="#"><i
+								class="fa fa-instagram" aria-hidden="true"></i></a></li>
 					</ul>
 				</div>
 			</div>
 		</div>
-		
+
 		<div class="copyright">
 			<div class="container">
 				<div class="row">
 					<div class="col-lg-12">
-						<p class="company-name">All Rights Reserved. &copy; 2018 <a href="#">Yamifood Restaurant</a> Design By : 
-					<a href="https://html.design/">html design</a></p>
+						<p class="company-name">
+							All Rights Reserved. &copy; 2018 <a href="#">Yamifood
+								Restaurant</a> Design By : <a href="https://html.design/">html
+								design</a>
+						</p>
 					</div>
 				</div>
 			</div>
 		</div>
-		
+
 	</footer>
 	<!-- End Footer -->
 </th:block>
